ref(roles): Refactor TeamSelect
